vantage-ai/requirements.txt
# Agent orchestration
google-adk>=1.0.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.27.0

# RAG / embeddings
openai>=1.30.0
sentence-transformers>=3.0.0
numpy>=1.26.0

# LiteLLM (ADK uses this to connect to Baseten)
litellm>=1.40.0

# Veris AI testing
veris>=0.1.0
